+++ eveuniverse/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "eveuniverse.apps.EveuniverseConfig"
 
-__version__ = "0.10.0"
+__version__ = "0.11.0"
 __title__ = "Eve Universe"